pyproject.toml
[project]
name = "markdownexporter"
version = "0.1.0"
description = "A dockerized solution that converts markdown files to PDF, Word, and HTML formats with MCP tools integration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markdown>=3.5.0",
    "python-markdown-math>=0.8",
    "weasyprint>=60.0",
    "reportlab>=4.0.0",
    "python-docx>=1.1.0",
    "docx2python>=2.9.0",
    "beautifulsoup4>=4.12.0",
    "jinja2>=3.1.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "structlog>=23.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "click>=8.1.0",
    "aiofiles>=23.0.0",
    "httpx>=0.25.0",
    "asyncio-mqtt>=0.16.0",
    "emoji>=2.8.0",
    "unidecode>=1.3.0",
    "pillow>=10.0.0",
    "requests>=2.31.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0"
]

[project.scripts]
markdownexporter = "markdownexporter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
asyncio_mode = "auto"
